+++ valhalla/config.h
@@ -6,7 +6,9 @@
 #define VALHALLA_STRINGIZE(A) VALHALLA_STRINGIZE_NX(A)
 
 /* Version number of package */
-#define VALHALLA_VERSION VALHALLA_STRINGIZE(VALHALLA_VERSION_MAJOR) "." VALHALLA_STRINGIZE(VALHALLA_VERSION_MINOR) "." VALHALLA_STRINGIZE(VALHALLA_VERSION_PATCH)
+#define VALHALLA_VERSION                                                                             \
+  VALHALLA_STRINGIZE(VALHALLA_VERSION_MAJOR)                                                         \
+  "." VALHALLA_STRINGIZE(VALHALLA_VERSION_MINOR) "." VALHALLA_STRINGIZE(VALHALLA_VERSION_PATCH)
 
 /* Name of package */
 #define PACKAGE "valhalla-" VALHALLA_VERSION
@@ -28,5 +30,3 @@
 
 /* Define to the version of this package. */
 #define PACKAGE_VERSION VALHALLA_VERSION
-
-